move tools/palette_h2.h to engine
because tools is linked against engine.a
