+++ fheroes2.vcxproj
@@ -383,6 +383,7 @@
     <ClInclude Include="src\engine\localevent.h" />
     <ClInclude Include="src\engine\math_base.h" />
     <ClInclude Include="src\engine\pal.h" />
+    <ClInclude Include="src\engine\palette_h2.h" />
     <ClInclude Include="src\engine\pathfinding.h" />
     <ClInclude Include="src\engine\rand.h" />
     <ClInclude Include="src\engine\rect.h" />
@@ -517,7 +518,6 @@
     <ClInclude Include="src\fheroes2\system\settings.h" />
     <ClInclude Include="src\fheroes2\world\world.h" />
     <ClInclude Include="src\fheroes2\world\world_pathfinding.h" />
-    <ClInclude Include="src\tools\palette_h2.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smacker.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smk_bitstream.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smk_hufftree.h" />
