+++ src/engine/pal.cpp
@@ -18,7 +18,6 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 #include "pal.h"
-#include "../../tools/palette_h2.h"
 
 #include <cstring>
